refactor(scoring): вынести scoring.py в пакет scoring/
Раскладка по модулям без изменения поведения (#491): types.py (общие типы),
employer.py (классификатор известности работодателя), prefilter.py (pre-LLM
фильтр #85), vacancy.py (эвристика + LLMScoringProvider #74). __init__.py
ре-экспортирует весь прежний публичный API scoring.py без изменений — ни один
импортирующий модуль (search.py, commands/_common.py, commands/search.py) и ни
один тестовый файл не изменён.

Фундамент для следующих видов скоринга (resume↔vacancy и letter↔vacancy
keyword-match), которые будут жить в этом же пакете.

Closes #491
